Move stuff around, cleanup imports
