+++ utils/objects-scraper/src/AbstractBase.py
@@ -1,0 +1,3 @@
+from sqlalchemy.orm import declarative_base
+
+Base = declarative_base()
